# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "M3th",
        "description": "math testing",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://node43.webte.fei.stuba.sk",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/questions": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "List questions, filterable by search string or tags",
                "operationId": "ccfcac18661863efc025fb5e958cb3bd",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in assignment_en or assignment_sk",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags[]",
                        "in": "query",
                        "description": "Filter by tag IDs (array)",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Questions fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "questions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Question"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Create a new question with answers and tags",
                "operationId": "3b45a42bdad6eb1d4eb6cdd88fbdd3d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assignment_sk",
                                    "assignment_en",
                                    "isMultiChoice",
                                    "answers"
                                ],
                                "properties": {
                                    "assignment_sk": {
                                        "type": "string",
                                        "example": "Koľko je 2+2?"
                                    },
                                    "assignment_en": {
                                        "type": "string",
                                        "example": "What is 2+2?"
                                    },
                                    "isMultiChoice": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "answer_sk": {
                                                    "type": "string",
                                                    "example": "Štyri"
                                                },
                                                "answer_en": {
                                                    "type": "string",
                                                    "example": "Four"
                                                },
                                                "isCorrect": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "tag_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Question"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/questions/{id}": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Show one question with answers and tags",
                "operationId": "d899529dc334e0fae06e945ff00bea1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the question",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question with answers and tags",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Question"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Questions"
                ],
                "summary": "Update a question, its answers, and tags",
                "operationId": "0d3414e242d85540e07f2f345fe85c62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the question",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assignment_sk",
                                    "assignment_en",
                                    "isMultiChoice",
                                    "tag_ids"
                                ],
                                "properties": {
                                    "assignment_sk": {
                                        "type": "string",
                                        "example": "Koľko je 2+2?"
                                    },
                                    "assignment_en": {
                                        "type": "string",
                                        "example": "What is 2+2?"
                                    },
                                    "isMultiChoice": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "answer_sk": {
                                                    "type": "string",
                                                    "example": "Štyri"
                                                },
                                                "answer_en": {
                                                    "type": "string",
                                                    "example": "Four"
                                                },
                                                "isCorrect": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "tag_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Question"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "Delete a question and its answers",
                "operationId": "12c5b45cc03b0643b3776bded072da33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the question",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Question 1 successfully deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Question": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "assignment_sk": {
                        "type": "string"
                    },
                    "assignment_en": {
                        "type": "string"
                    },
                    "isMultiChoice": {
                        "type": "boolean"
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Answer"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    }
                },
                "type": "object"
            },
            "Answer": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "question_id": {
                        "type": "integer"
                    },
                    "answer_sk": {
                        "type": "string"
                    },
                    "answer_en": {
                        "type": "string"
                    },
                    "isCorrect": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name_sk": {
                        "type": "string"
                    },
                    "name_en": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}